Machine Learning Final Project
Pump It Up is an online competition which was chosen as a part of my
Machine Learning course. Using data from Taarifa and the Tanzanian
Ministry of Water, prediction of pumps which are functional, which need
some repairs, and which don't work at all should be made. The
prediction is based on a number of variables. The classifier used is
Random Forests.